@@ -23,7 +23,7 @@ license = {file = "LICENSE"}
 authors = [
     { name = "meejah", email = "meejah@meejah.ca" },
 ]
-requires-python = ">=3.6"
+requires-python = ">=3.9"
 keywords = [
     "cryptography",
     "forwarding",
@@ -43,7 +43,7 @@ dependencies = [
     "msgpack",
     "humanize",
     "twisted",
-    "magic-wormhole[dilate] @ git+https://github.com/magic-wormhole/magic-wormhole.git#egg=magic-wormhole[dilate]",
+    "magic-wormhole[dilate]",
     "rich",
 ]
 
